Fair Visualizations 6 - Added other and missing counts (#267)
* fiar vis 6 - Added other and missing counts

* fair vis 6 - moved color palettes to separate file

* Removed unnecessary code

* removed unused import

* Rewrite sitemap url to external s3 bucket

* Fix rewrite for sitemap

Should be /sitemaps/path

* Fix cypress test requiring sitemaps url

* bump version of cypress action

* Use inherited secrets

* Specifiy node version to test under

* Explicitly set node-version

* Fixed color palette bug

* Cosmetic fixes

* Fixed typo

* Re-aligned graph titles

---------
