+++ src/components/DonutChart/DonutChart.tsx
@@ -27,272 +27,6 @@ const actions = {
   editor: false
 }
 
-export const otherRange = [ 'gray', 'black' ]
-export const otherDomain = [ 'Other', 'Missing' ]
-
-// Source:  https://r-charts.com/color-palettes/#discrete
-// paletteer_d("ggsci::category20_d3") + some from original set
-export const typesRange = [
-  '#1F77B4FF',
-  '#FF7F0EFF',
-  '#2CA02CFF',
-  '#D62728FF',
-  '#9467BDFF',
-  '#8C564BFF',
-  '#E377C2FF',
-  '#7F7F7FFF',
-  '#BCBD22FF',
-  '#17BECFFF',
-  '#AEC7E8FF',
-  '#FFBB78FF',
-  '#98DF8AFF',
-  '#FF9896FF',
-  '#C5B0D5FF',
-  '#C49C94FF',
-  '#F7B6D2FF',
-  '#C7C7C7FF',
-  '#DBDB8DFF',
-  '#9EDAE5FF',
-  '#fccde5',
-  '#fdb462',
-  '#fb8072',
-  '#b3de69',
-  '#bc80bd',
-  '#fccde5',
-  '#8dd3c7',
-  ...otherRange
-]
-// Source: https://schema.datacite.org/meta/kernel-4/include/datacite-resourceType-v4.xsd
-export const typesDomain = [
-  'Audiovisual',
-  'Book',
-  'Book Chapter',
-  'Collection',
-  'Computational Notebook',
-  'Conference Paper',
-  'Conference Proceeding',
-  'Data Paper',
-  'Dataset',
-  'Dissertation',
-  'Event',
-  'Image',
-  'Interactive Resource',
-  'Journal',
-  'Journal Article',
-  'Model',
-  'Output Management Plan',
-  'Peer Review',
-  'Physical Object',
-  'Preprint',
-  'Report',
-  'Service',
-  'Software',
-  'Sound',
-  'Standard',
-  'Text',
-  'Workflow',
-  ...otherDomain
-]
-
-export const contributorRange = [
-  'red',
-  'green',
-  'blue'
-]
-
-export const contributorDomain = [
-  'Creator',
-  'Data Curator',
-  'Project Leader'
-]
-
-export const affiliationRange = [
-  'red',
-  'green',
-  'blue'
-]
-
-export const affiliationDomain = [
-  'Affiliation 1',
-  'Affiliation 2',
-  'Affiliation 3'
-]
-
-
-export const licenseRange = [
-  '#d9d9d9',
-  '#ffffb3',
-  '#ccebc5',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#8dd3c7',
-  '#80b1d3',
-  '#b3de69',
-  '#fdb462',
-  '#fccde5',
-  '#ccebc5',
-  '#ccebc5',
-  '#ffed6f',
-  ...otherRange
-]
-
-export const licenseDomain = [
-  'No License',
-  'AFL-1.1',
-  'Apache-2.0',
-  'CC-BY-1.0',
-  'CC-BY-2.0',
-  'CC-BY-2.5',
-  'CC-BY-3.0',
-  'CC-BY-4.0',
-  'CC-BY-NC-2.0',
-  'CC-BY-NC-2.5',
-  'CC-BY-NC-3.0',
-  'CC-BY-NC-4.0',
-  'CC-BY-NC-ND-3.0',
-  'CC-BY-NC-ND-4.0',
-  'CC-BY-NC-SA-3.0',
-  'CC-BY-NC-SA-4.0',
-  'CC-BY-SA-4.0',
-  'CC-PDDC',
-  'CC0-1.0',
-  'GPL-3.0',
-  'ISC',
-  'MIT',
-  'MPL-2.0',
-  'OGL-Canada-2.0',
-  ...otherDomain
-]
-
-export const fieldOfScienceRange = [
-  '#d9d9d9',
-  '#ffffb3',
-  '#ccebc5',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#8dd3c7',
-  '#80b1d3',
-  '#b3de69',
-  '#fdb462',
-  '#fccde5',
-  '#ccebc5',
-  '#ccebc5',
-  '#ffed6f'
-]
-
-export const fieldOfScienceDomain = [
-  'No Field of Science',
-  'Natural sciences',
-  'Mathematics',
-  'Computer and information sciences',
-  'Physical sciences',
-  'Chemical sciences',
-  'Earth and related environmental sciences',
-  'Biological sciences',
-  'Other natural sciences',
-  'Engineering and technology',
-  'Civil engineering',
-  'Electrical engineering, electronic engineering, information engineering',
-  'Mechanical engineering',
-  'Chemical engineering',
-  'Materials engineering',
-  'Medical engineering',
-  'Environmental engineering',
-  'Environmental biotechnology',
-  'Industrial biotechnology',
-  'Nano-technology',
-  'Other engineering and technologies',
-  'Medical and health sciences',
-  'Basic medicine',
-  'Clinical medicine',
-  'Health sciences',
-  'Medical biotechnology',
-  'Other medical sciences',
-  'Agricultural sciences',
-  'Agriculture, forestry, and fisheries',
-  'Animal and dairy science',
-  'Veterinary science',
-  'Agricultural biotechnology',
-  'Other agricultural sciences',
-  'Social sciences',
-  'Psychology',
-  'Economics and business',
-  'Educational sciences',
-  'Sociology',
-  'Law',
-  'Political science',
-  'Social and economic geography',
-  'Media and communications',
-  'Other social sciences',
-  'Humanities',
-  'History and archaeology',
-  'Languages and literature',
-  'Philosophy, ethics and religion',
-  'Arts (arts, history of arts, performing arts, music)',
-  'Other humanities'
-]
-
-export const registrationAgencyRange = [
-  '#d9d9d9',
-  '#ffffb3',
-  '#ccebc5',
-  '#E81A31',
-  '#159DEA',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#8dd3c7',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#80b1d3',
-  '#8dd3c7',
-  '#80b1d3',
-  '#b3de69',
-  '#fdb462',
-  '#fccde5',
-  '#ccebc5',
-  '#ccebc5',
-  '#ffed6f'
-]
-
-export const registrationAgencyDomain = [
-  'No Registration Agency',
-  'Airiti',
-  'CNKI',
-  'Crossref',
-  'DataCite',
-  'ISTIC',
-  'JaLC',
-  'KISTI',
-  'mEDRA',
-  'OP'
-]
 
 const DonutChart: React.FunctionComponent<Props> = ({
   data,
