+++ src/components/WorksDashboard/WorksDashboard.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 import { Row, Col } from 'react-bootstrap'
 import clone from 'lodash/clone'
-import { ContentFacet } from '../WorksListing/WorksListing'
 import { Works } from '../SearchWork/SearchWork'
-import { typesDomain, typesRange, licenseDomain, licenseRange } from '../DonutChart/DonutChart'
 import ProductionChart from '../ProductionChart/ProductionChart'
 import HorizontalStackedBarChart, { getTopFive, toBarRecord } from '../HorizontalStackedBarChart/HorizontalStackedBarChart'
+import { resourceTypeDomain, resourceTypeRange, licenseRange, identifierDomain, identifierRange, otherDomain, otherRange } from '../../data/color_palettes'
 
 type Props = {
   works: Works
 }
 
+const tooltipText = (sourceField: string) => `The field "${sourceField}" from DOI metadata was used to generate this chart.`
+
 const WorksDashboard: React.FunctionComponent<Props> = ({ works, children }) => {
   if (works.totalCount == 0) return null
-  // const hasNoWorks = works.totalCount == 0
 
   const published = works.published.map((x) => ({
     title: x.title,
@@ -21,17 +21,7 @@ const WorksDashboard: React.FunctionComponent<Props> = ({ works, children }) => 
   }))
 
   const resourceTypes = getTopFive(works.resourceTypes.map(toBarRecord))
-  
-
-  const noLicenseValue: ContentFacet = {
-    id: 'no-license',
-    title: 'No License',
-    count:
-      works.totalCount -
-      works.licenses.reduce((a, b) => a + (b['count'] || 0), 0)
-  }
   const licensesData = clone(works.licenses)
-  licensesData.unshift(noLicenseValue)
   const licenses = getTopFive(licensesData.map(toBarRecord))
 
   return (
@@ -49,27 +39,29 @@ const WorksDashboard: React.FunctionComponent<Props> = ({ works, children }) => 
       <Row>
         <Col xs={12} sm={4}>
           <HorizontalStackedBarChart
-            titlePercent={-1}
-            titleText={[`of scholarly outputs use`, `a persistent identifier (i.e. DOI)`]}
-            data={[{title: 'PLACEHOLDER', count: 0}]}
-            domain={[]}
-            range={[]} />
+            titlePercent={100}
+            titleText={['of scholarly outputs use', 'a persistent identifier (i.e. DOI)']}
+            data={[{title: 'DOI', count: 1}]}
+            domain={identifierDomain}
+            range={identifierRange} />
         </Col>
         <Col xs={12} sm={4}>
           <HorizontalStackedBarChart
             titlePercent={resourceTypes.topPercent}
-            titleText={`of scholarly outputs are ${resourceTypes.topCategory}`}
+            titleText={[`of scholarly outputs are ${resourceTypes.topCategory}`, '']}
             data={resourceTypes.data}
-            domain={typesDomain}
-            range={typesRange} />
+            domain={resourceTypeDomain}
+            range={resourceTypeRange}
+            tooltipText={tooltipText('resourceTypes')} />
         </Col>
         <Col xs={12} sm={4}>
           <HorizontalStackedBarChart 
             titlePercent={licenses.topPercent}
-            titleText={`of scholarly outputs use ${licenses.topCategory}`}
+            titleText={[`of scholarly outputs use ${licenses.topCategory}`, '']}
             data={licenses.data}
-            domain={licenseDomain}
-            range={licenseRange} />
+            domain={[...otherDomain, ...licenses.data.map(l => l.title)]}
+            range={[...otherRange, ...licenseRange]}
+            tooltipText={tooltipText('licenses')} />
         </Col>
       </Row>
     </>
